Save ProductAttributes and others

+++ src/app/views/pages/freniteca/masters/products-createupdate/products-createupdate.component.ts
@@ -1,4 +1,3 @@
-import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { FormArray, FormBuilder, Validators } from '@angular/forms';
 import { DomSanitizer } from '@angular/platform-browser';
@@ -11,6 +10,7 @@ import { ProductAttributeModel } from 'src/app/models/productAttribute.model';
 import { TypeProductModel } from 'src/app/models/typeProduct.model';
 import { TypeProductAttributeModel } from 'src/app/models/typeProductAttribute.model';
 import { ApiService } from 'src/app/services/api.service';
+import { ProductAttributeService } from 'src/app/services/product-attribute.service';
 import { ProductService } from 'src/app/services/product.service';
 import { environment } from 'src/environments/environment';
 import Swal from 'sweetalert2';
@@ -52,27 +52,27 @@ export class ProductsCreateupdateComponent implements OnInit {
   typeProducts: TypeProductModel[]=[];
   Applications: ApplicationModel[] = [];
   brands: BrandModel[]=[];
-  productsEquivalence: ProductModel[]=[];
   typeProductsAttributes: TypeProductAttributeModel[] = [];
   urlImages: string = '';
   listImagenes: string[] = [];
+  isView: boolean= false;
 
   constructor(private route: ActivatedRoute, 
     private api: ApiService, 
     private router: Router, 
     private fb: FormBuilder,
     private sanitizer: DomSanitizer, 
+    private productAttributeService: ProductAttributeService,
     private productService: ProductService) {
-   // this.detail.push(this.frmDetail);
-   //this.addDetail();
    this.urlImages = environment.urlImages;
    }
 
   async ngOnInit(): Promise<void> {
     this.id = this.route.snapshot.paramMap.get('id')??'';
+    this.isView = this.route.snapshot.paramMap.get('view')=='view'??false;
+
     await this.getTypeProducts();
     await this.getBrands();
-    await this.getProductsEquivalence();
     await this.getApplications();
 
     if (this.id === '') {
@@ -80,7 +80,16 @@ export class ProductsCreateupdateComponent implements OnInit {
       this.reg = new ProductModel();
       await this.setFields();
     } else {
-      this.subtitle = 'EDITANDO';
+      if (this.isView)
+      {
+        this.subtitle = 'CONSULTANDO';
+        this.disabledAll();
+      }
+      else
+      {
+        this.subtitle = 'EDITANDO';
+
+      }
       await this.api.getId('product',this.id).subscribe(
         async (resp: any) => {
           this.reg = resp.data;
@@ -117,27 +126,19 @@ export class ProductsCreateupdateComponent implements OnInit {
       
                   }
                 );
+            }
+          );
 
-                this.api.getCustom('productApplication','listbyProductId','ProductId',this.id).subscribe(
-                  (resp: any) => {
-                    this.ProductAttribute.clear();
-                    if (resp.status) {
+          await this.api.getCustom('productApplication','listbyProductId','ProductId',this.id).subscribe(
+            (resp: any) => {
+              this.ProductAttribute.clear();
+              if (resp.status) {
+                
+                this.ProductApplications = resp.data;
+                this.loadProductApplications();
+              }
 
-                      this.ProductApplications = resp.data;
-                      this.typeProductsAttributes.forEach((element: TypeProductAttributeModel)=>{
-                        var res = this.ProductAttributes.filter(x=>x.typeProductAttributeId == element.id);
-                        if (res.length >0)
-                        {
-                          element.value = res[0].value;
-                        }
-                      });              
-  
-                      this.loadProductAttributes();
-                    }
-      
-      
-                  }
-                );
+
             }
           );
           
@@ -150,25 +151,50 @@ export class ProductsCreateupdateComponent implements OnInit {
   async getTypeProducts(){
     this.api.get('typeProduct').subscribe(
       (resp: any) => {
-        if (resp.status)
+        if (resp.status) {
+          resp.data.unshift(
+            {
+              id: '',
+              code: '',
+              description: 'Seleccione un tipo de producto',
+              dateCreation: new Date(),
+              dateLastUpdate: new Date(),
+              status: 'active'
+            }
+          );
           this.typeProducts = resp.data;
+        }
       }
     );
   }
-
+  disabledAll(){
+    this.frm.controls['code'].disable();
+    this.frm.controls['description'].disable();
+    this.frm.controls['reference'].disable();
+    this.frm.controls['referenceProvider'].disable();
+    this.frm.controls['brandId'].disable();
+    this.frm.controls['typeProductId'].disable();
+    this.frm.controls['typeProductId'].disable();
+    this.frm.controls['productEquivalenceId'].disable();
+      
+    
+  }
   async getBrands(){
     this.api.get('brand').subscribe(
       (resp: any) => {
-        if (resp.status)
+        if (resp.status) {
+          resp.data.unshift(
+            {
+              id: '',
+              code: '',
+              description: 'Seleccione una marca',
+              dateCreation: new Date(),
+              dateLastUpdate: new Date(),
+              status: 'active'
+            }
+          );
           this.brands = resp.data;
-      }
-    );
-  }
-  async getProductsEquivalence(){
-    this.api.get('product').subscribe(
-      (resp: any) => {
-        if (resp.status)
-          this.productsEquivalence = resp.data;
+        }
       }
     );
   }
@@ -176,8 +202,19 @@ export class ProductsCreateupdateComponent implements OnInit {
     this.api.get('application').subscribe(
       (resp: any) => {
         if (resp.status)
+        {
+          resp.data.unshift(
+            {
+              id: '',
+              code: '',
+              description: 'Seleccione una aplicación',
+              dateCreation: new Date(),
+              dateLastUpdate: new Date(),
+              status: 'active'
+            }
+          );
           this.Applications = resp.data;
-        console.log('this.Applications', this.Applications);
+        }
       }
     );
   }
@@ -217,10 +254,25 @@ export class ProductsCreateupdateComponent implements OnInit {
             if (resp.error) {
                 Swal.fire('Error al crear el Registro','Se presentó un error al crear el registro', 'error');
             } else {
+              
 //TODO: Guardar datos de atributos
+            var controls = this.ProductAttribute.controls;
+            controls.forEach((element:any)=>{
+              let productAttributeModel: ProductAttributeModel = {
+                id: uuidv4(),
+                productId : this.reg.id,
+                typeProductAttributeId : element.value.id,
+                value : element.value.value,
+                dateCreation: new Date(),
+                status: 'active',
+                dateLastUpdate: new Date()                
+              };
+              this.api.create('ProductAttribute',productAttributeModel).subscribe();
+            });
+
 //TODO: Guardar documentos
 //TODO: Guardar Aplicaciones
-              this.router.navigateByUrl('/masters/products');
+              this.router.navigateByUrl('/masters/products/admin');
             }
           });
         } else {
@@ -229,10 +281,55 @@ export class ProductsCreateupdateComponent implements OnInit {
             if (resp.error) {
                 Swal.fire('Error al actualizar el Registro','Se presentó un error al actualizar el registro', 'error');
             } else {
-//TODO: Actualizar datos de atributos
+              var controls = this.ProductAttribute.controls;
+              controls.forEach((element:any)=>{
+                this.productAttributeService.getByProductIdAndTypeProductAttributeId(this.id, element.value.id).subscribe(
+                  (resp: any) => {
+                    if (resp.status) {
+                      let productAttributeModel: ProductAttributeModel = {
+                        id: '',
+                        productId : this.reg.id,
+                        typeProductAttributeId : element.value.id,
+                        value : element.value.value,
+                        dateCreation: new Date(),
+                        status: 'active',
+                        dateLastUpdate: new Date()                
+                      };
+                      if (resp.data == null) // crear registros
+                      {
+                        productAttributeModel.id = uuidv4();
+                        this.api.create('productAttribute', productAttributeModel).subscribe(
+                          (resp:any) => {
+                            if (resp.status){
+                              console.log('productAttribute Creado');
+                            }
+                          },
+                          (error: any) => { console.log('error Creado', productAttributeModel); }
+                        );
+                      } 
+                      else {
+                        productAttributeModel.id = resp.data.id;
+                        productAttributeModel.dateCreation = resp.data.dateCreation;
+                        this.api.update('productAttribute', productAttributeModel).subscribe(
+                          (resp:any) => {
+                            if (resp.status){
+                              console.log('productAttribute Creado');
+                            }
+                          },
+                          (error: any) => { console.log('error Creado', productAttributeModel); }
+                        );
+
+                      }
+                    
+                      console.log('resp', resp);
+                    }
+                  }
+                );
+              });
+
 //TODO: Guardar documentos
 //TODO: Guardar Aplicaciones
-              this.router.navigateByUrl('/masters/products');
+              this.router.navigateByUrl('/masters/products/admin');
             }
           });
         }
@@ -247,7 +344,6 @@ export class ProductsCreateupdateComponent implements OnInit {
     this.reg.description = this.frm.get('description')?.value??'';
     this.reg.brandId = this.frm.get('brandId')?.value??'';
     this.reg.typeProductId = this.frm.get('typeProductId')?.value ?? '';
-    this.reg.productEquivalenceId = this.frm.get('productEquivalence')?.value ?? '';
 
 
   }
@@ -258,8 +354,9 @@ export class ProductsCreateupdateComponent implements OnInit {
     this.frm.controls['referenceProvider'].setValue(this.reg.referenceProvider);
     this.frm.controls['brandId'].setValue(this.reg.brandId);
     this.frm.controls['typeProductId'].setValue(this.reg.typeProductId);
-    this.frm.controls['typeProductId'].disable();
-    this.frm.controls['productEquivalenceId'].setValue(this.reg.productEquivalenceId=='00000000-0000-0000-0000-000000000000' ? '' : this.reg.productEquivalenceId);
+    if (this.id !== ''){
+      this.frm.controls['typeProductId'].disable();
+    }
     await this.selectedTypeProduct(this.reg.typeProductId);
     
     
@@ -270,7 +367,7 @@ export class ProductsCreateupdateComponent implements OnInit {
   }
   addTypeProductAttribute(typeProductId:string = '', descriptionTypeProduct: string = '', value: string = '', id: string = ''){
    const  frmDetail = this.fb.group({
-      typeProductAttributeId: [typeProductId],
+      typeProductId: [typeProductId],
       descriptionTypeProduct: [descriptionTypeProduct],
       value: [value, Validators.required],
       id: [id],
@@ -298,34 +395,35 @@ export class ProductsCreateupdateComponent implements OnInit {
   }
   deleteApplication(idx: number){
     var reg = this.ProductApplication.at(idx);
-
-    this.api.delete('ProductApplication',reg.value.id).subscribe(
-      (resp: any) => {
-        if (resp.status)
-          this.ProductApplication.removeAt(idx);
-          this.Applications.push(reg.value['application']);
-      }
-    );
+    if (reg.value.id != '')
+    {
+      this.api.delete('ProductApplication',reg.value.id).subscribe(
+        (resp: any) => {
+          if (resp.status)
+          {
+            this.Applications.push(reg.value['application']);
+          }
+          
+        }
+        );
+      } 
+      this.ProductApplication.removeAt(idx);
 
   }
   selectedApplication(idx: number, id: string, application: ApplicationModel){
+    debugger;
     var reg = this.ProductApplication.at(idx);
-    var regEncontrar = this.Applications.findIndex(x=>x.id == id);
     reg.value['applicationId'] = id;
     reg.value['id'] = this.id;
+    var regEncontrar = this.Applications.findIndex(x=>x.id == id);
     reg.value['application'] = this.Applications[regEncontrar];
     this.Applications.splice(regEncontrar, 1);
     //this.loadProductApplications();
   }
   loadProductApplications(){
     this.ProductApplications.forEach((element:any)=>{
-      this.api.getId('Application', element.applicationId).subscribe(
-        (resp:any) => {
-          if (resp.status){
-            this.addApplication(element.applicationId, element.id, resp.data);
-          }
-        }
-      );
+      this.addApplication(element.applicationId, element.applicationId, element.applicationNavigation);
+
     });
   }
   async selectedTypeProduct(id:string){
@@ -377,5 +475,14 @@ export class ProductsCreateupdateComponent implements OnInit {
   transform(url: string) {
     return this.sanitizer.bypassSecurityTrustResourceUrl(url);
   }
-  
+  back(){
+    if (this.isView){
+      this.router.navigateByUrl('/masters/products/ppal');
+
+    }else
+    {
+      this.router.navigateByUrl('/masters/products/admin');
+
+    }
+  }
 }
